Add 'episode' mediatype to episode items
Sometimes the plot/description text for an episode is too big for the UI space on Kodi and becomes truncated.  
Setting the mediatype infolabel of episode items to value 'episode' lets them show the 'Information' context-menu item, which leads you to a Kodi screen where you can read the plot text in full.